Factor out create job success job creation
